Python/requirements.txt
# Computation
pyvista
imageio
numpy
PyQt5
scipy
pandas
pyyaml
dash
imagio
imageio-ffmpeg

# Testing
pytest-cov
pytest-html
pytest-subtests
mypy
types-PyYAML

# Traceability
tzlocal

# Custom
git+https://github.com/LonghornRacingElectric/tire_analysis@LHR_tire_toolkit_package
